+++ packages/sprinkle-core/app/src/ServicesProvider/MarkdownService.php
@@ -25,6 +25,8 @@ use UserFrosting\ServicesProvider\ServicesProviderInterface;
  * Markdown service. Add CommonMark markdown parser with GitHub Flavored Markdown frontmatter support.
  *
  * @see https://commonmark.thephpleague.com
+ *
+ * TODO : Should have a way to extend the markdown parser with custom extensions.
  */
 class MarkdownService implements ServicesProviderInterface
 {
@@ -32,7 +34,10 @@ class MarkdownService implements ServicesProviderInterface
     {
         return [
             ConverterInterface::class => function (Config $config) {
-                $environment = new Environment([]);
+                // Get markdown configuration from config service
+                $markdownConfig = $config->get('markdown', []);
+
+                $environment = new Environment($markdownConfig);
                 $environment->addExtension(new CommonMarkCoreExtension());
                 $environment->addExtension(new FrontMatterExtension());
                 $environment->addExtension(new GithubFlavoredMarkdownExtension());